fix(filter): stop to list FS to exclude files to scan
ggshield supports to exclude some files from scan using exclude
patterns.  Those patterns can be set using cli arguments, configuration
file and there are default excluded patterns as well.

We used to list all the files matching those patterns using glob in
order to filter out files to scan. We recently added some default
patterns to ignore and it leds us to realize the performance issue with
this approach: FS listing is very slow.

We now use regex to check if a file path is included in the exclude
patterns.