fix(docker): fix "docker save" failing on Windows (#161)
Saving fails if the image name contains a ':' character because we then
try to create a tarball name with a ':' in it, which is not an allowed
filename character on Windows.

The tarball is created in a temporary directory, so it does not need to
look like the image name. Let's use a fixed name instead: "archive.tar".

This commit also uses Path instead of str as a type for the path arguments
of functions:
- docker_scan_archive
- docker_archive_cmd
- docker_save_to_tmp
- get_files_from_docker_archive