fix: fix pre-receive hook error handling with GitLab Web UI
GitLab WebUI filters the pre-receive output and only shows lines
starting with GL-HOOK-ERR.

We now detect if the pre-receive hook is running from GitLab Web UI. If
it's the case we use a shorter message, prefixed with GL-HOOK-ERR so that
the message shows up in the UI.

We no longer disable the hook completely when running from the Web UI,
so the --web option is no longer useful. As such it is now deprecated.

Fixes #151.