fix(docker): fix crash when scanning a non-local image
If we run `ggshield scan docker <image>` and `<image>` is not available
locally, `ggshield` downloads it first.

I broke this feature when I made other Docker fixes: ggshield currently
raises an exception at the end of the download.

Fix that and add tests to cover that case to ensure that cannot go
undetected anymore.

Fixes #181.