fix(packaging): make ggshield package work even if `python3` is too old
Wrap call to ggshield in a shell script which tries to find the most
appropriate version of Python. This makes the rpm package work on systems
like RedHat redhat/ubi8 Docker images, where `python3` is 3.6 but there are
python 3.8 packages available.

A better solution would be for the package to depend on the python3.8
package but we can't do that unless we decide to build one package per
distribution version. We don't have the bandwidth to do that for now.