fix(config): do not store "config-path" key when saving user file
Remove YAMLFileConfig save() and load() abstract methods because
now that UserConfig.save() signature does not match YAMLFileConfig.save()
anymore, mypy is unhappy. Having a YAMLFileConfig.save() method would be
useful if some of our code received a UserConfig instance as a
YAMLFileConfig instance and expected to be able to call save() on it, but
it's not the case anyway, so we can remove it. For consistency, also remove
YAMLFileConfig.load().