fix(config): store `instance`, do not store `dashboard-url`
Remove `api_url` and `dashboard_url` from UserConfig. Replace them with
`instance`.

Refactor UserConfig and AuthConfig to only contain what they store in
config files. This means the following changes:
- If the user overrides the default instance using the command-line, the
  instance name is stored in `UserConfig._cmdline_instance_name`.
- UserConfig and AuthConfig no longer care about environment variables.
  Config is taking care about them.

Replace GITGUARDIAN_URL environment variable with GITGUARDIAN_INSTANCE,
so that the code is ready for when a name can be associated to an instance.