fix: fix diff parser failing on some patches
Parser would identify a permission change + rename + edit as a
permission change. It would thus fail to parse the filename because in
this case the `diff --git a/<old> b/<new>` has different values for
`old` and `new`.

This commit rewrites the parser to use the `-z --raw` format, which
outputs the affected filenames in a machine-friendly header. This header
is parsed before extracting the diffs.