fix(patch-parser): fix parsing of merge commits
The changes from #327 were not so good because while ggshield was able to
parse merge commits and weird file names, if it found a secrets in a merge
commit it would fail to parse the diff. This is because a merge commit
diff has a slightly different format than a single-parent commit: it
contains one column per parent for its status for example.

To fix that bug, this commits does the following:
- call `git show` and `git diff` with `-m`. This makes git
  generate multiple single-parent commits for merge commits.
- Adjust the patch parser to take into account all commits returned by the
  command.

With these changes we are back to only single-parent commits, which
ggshield knows how to handle.