feat(scan_repo): optimize scan repo command
The `ggshield secret scan repo` command used to trigger one API call per commit.
This implementation was not efficient. We change this to group commits together
as long as the total number of files sent to the API is below the fixed limit.

The implementation modifies scan_commit_range in accordance. We also had to adapt
the results parsing. Now that one API call contains files from several commits,
we have to re-create the link between the results and the commit they come from.

One small drawback is that we lose this commmit granularity when it comes to error :
we cannot tell which commit triggered which error.