feat(scan path): do not read all files at startup
This makes ggshield respond faster and reduces memory usage when scanning
large file trees.

Not loading all files at startup is important because when we start to
detect the file encoding, reading files is going to be slower, so it
must not happen before we ask the user to confirm they want to run the
scan.