fix(scan repo): do not ignore failures when run from a git repo
If the current working directory is an existing git checkout, then
running `secret scan repo SOMETHING` does not produce an error if
`SOMETHING` is neither the URL of a git repository or a valid path to a
git checkout.

```
$ ggshield secret scan repo https://example.com/doesnotexist.git
Error: /tmp/tmpaiyn0u5r is not a git repository

$ mkdir foo

$ cd foo

$ git init

$ ggshield secret scan repo https://example.com/doesnotexist.git
Scanning Commits  [####################################]  100%

No secrets have been found
```

This is because we cache the return value of `check_git_dir(wd)` but
this function can be called with `wd = None` in which case it uses
`os.getcwd()` as the path. If `os.getcwd()` changes between two
`check_git_dir()` calls then the returned value for the second call is
wrong.

To fix this we:
- make `is_git_dir()` do the bulk of the work, remove its default value
and cache its results
- remove the caching decorator from `check_git_dir()`

(Doing it this way was less intrusive than removing the default
value of `check_git_dir()`)