fix(pre-commit): fix `pre-commit` calling ggshield multiple times
By default, the pre-commit framework calls hooks with files to scan as
arguments. If the file list is too long, it splits it and run scans in
parallel.

This is not adapted for ggshield because `ggshield secret scan pre-commit`
(and `pre-push`) gets the content to scan from the commit ID, the arguments
are ignored. When pre-commit calls ggshield twice, ggshield scans the whole
commit twice.

By telling pre-commit to not call ggshield with filenames, we avoid this
issue: ggshield is only called once. The loss of parallelism is not an
problem since ggshield handles parallelism itself.