feat(check_for_updates): improve the check_for_updates option
We make `--check-for-option/--no-check-for-option` a common option.
The value for the flag is store in ctx.obj, we change the default value for the option to None to avoid
overwriting the value in the context when walking down subcommands. We then handle this None to fallback to True.
The execution of check_for_updates is moved out of cli, this is in order to log in accordance with the debug option.