fix: fix "dubious dir" errors when used with GitLab
When GitLab clones a repository inside the container, it does so as
root, in a directory which is not $PWD. When it then runs command inside
the container, the commands are run as our `app` user. When ggshield
runs a git command in the checkout directory it fails because files are
owned by root and the directory has not been marked as safe.

Since we can't know the directory in advance and since all CI servers I
know of expect the default user in the image to be root (GitHub actively
discourages using another user in Dockerfile-based GitHub actions, see
[1]), cave in and remove our `app` user.

[1]: https://docs.github.com/en/actions/creating-actions/dockerfile-support-for-github-actions#user