chore(deps): # Clean Pipfile
We used to list dependencies of ggshield both in the Pipfile and in the setup.py
This could lead to discrepancies between the two files, and hence packages being not correctly shipped with gg-shield.
To prevent this in the future, this PR gets rid of gg-shield dependencies in the Pipfile.
Pipenv will install the package along with its dependencies anyway.