fix: make `ggshield auth login` handle server-side removed token
If the token has been removed on the server-side but is still in ggshield
config, calling `ggshield auth login` would say everything is fine, but
the next call to scan would fail.

To fix this, make `ggshield auth login` check the stored token is valid,
and if it's not, forget it and restart the login process.