chore(ci): fix `publish-gh-release` failure
When generating the changelog heading for the new version, scriv uses an
em dash (—) to separate the version from the date, not a hyphen dash. Do
not try to match it when extracting the notes for the release.

Update the existing entry in CHANGELOG.md to use an em dash.