fix: use DiffID instead of the random ID used to store layer tarballs
Layers have a DiffID computed from their content. This is a more
interesting ID for caching than the random ID used as a folder name for
layers, because we know a layer DiffID only changes when its content
changes.

As a side-effect, remove LayerInfo.get_id(): it's no longer used.