chore(ci): use `gh` to create release instead of actions
The initial reason for this change is that `actions/create-release` is no
longer maintained and has started generating deprecation warnings about
set-output usage. Switch to using `gh` since this is maintained by GitHub.

While we are at it, use `gh` to upload assets too. Using `gh` instead of
actions makes it easier to run steps manually if something goes wrong.