feat: add --log-file option, rework logging
`--log-file` redirects log output to a log file, making it easier to
process.

No log messages are now printed by default, even logging.ERROR. This avoids
mixing logging-style output with regular CLI output. This should not be a
problem because every time we log a message with ERROR level, we also
display it using display_error().

Additionally, the option can be activated using the
`GITGUARDIAN_LOG_FILE` environment variable. This makes it possible to
enable logs in situations where it's not easy to modify arguments, such
as pre-commit.