feat(oauth): #COR-1425 Support max lifetime on access tokens
A new maximum lifetime setting is coming for business workspaces. With
this setting, managers can set a maximum lifetime for personal access
tokens.

When a token is created with a lifetime above this limit, its expiration
date is downsized to respect the workspace setting. THen, a warning
message is added to the login output to inform the user.

Changelog: skip