fix: make Scannable.is_longer_than() compare to the correct size
`Scannable.is_longer_than()` currently compares the size passed to it to
the decoded size, but what we actually want to compare it to is the
*UTF-8 encoded size*.

To implement this without reading the content multiple times,
`Scannable` implementations now store the UTF-8 encoded size, and
`Scannable` helper methods (`_is_file_longer_than()` and
`_decode_bytes()`) return the UTF-8 encoded size in addition to their
current return value.

When the content they work on is already UTF-8, the helper methods try
to avoid reading it until necessary.