fix: file content could not be found
In case of a scan launched on a directory within a repository the
related file content was not found.

The base directory name used was the path provided and the vulnerability
file name was relative to the repository root which lead to wrong file
path.

The path provided is not used as is anymore and is used to find the
repository root.

Issue: IAC-246
Changelog: skip