feat(oauth): improve error handling
When the OAuth process fails, show the error message provided in the
response if it is there. This way the user can understand why the process
failed (for example because they requested a scope they are not allowed to
request).

Additionally, in the browser, show a basic HTML page with the error
instead of a blank page.