feat: Introduce GGShieldUI
The goal here is to have all progress-feedback code go through GGShieldUI
so that messages and progress bars stop stepping on each others.

Depending on whether stderr is a tty or not, ggshield uses either
RichGGShieldUI or PlainTextGGShieldUI.

SecretScannerUI has been turned into a more generic ScannerUI so that other
verticals can make use of it to report scanning progress.

All progress bars are now created through GGShieldUI. When ScannerUI is not
adapted, for example when showing progress on a group of scans, then the
code use a more generic progress bar: GGShieldProgress.