feat(perfbench): make `run` `-V` option accept a path to a ggshield exe
This means versions can now be random strings, so the order in which
versions are passed to `perfbench run` matters: the first version is the
reference version, against which others are compared.

This requires storing the version order in the raw report file generated
by `perfbench run`. To do so the raw report file becomes a JSON instead
of a JSONL, containing an object with two keys:

- `versions`: a list of versions
- `entries`: a list of run entries (what was previously in the JSONL)