fix: fix configuration override
If a key is set in the global configuration file, but not in the local one,
then the global value is ignored unless it's a list key.

The current code loads each configuration file (global and local) as
UserConfig instances, then merge the instances.

When the local UserConfig instance is loaded, if a key is not set, its
field uses the default value. This is a problem because then the code
responsible for merging the global and local UserConfig instances considers
that the local value has explicitly been set and ignores the global value.

With this change, ggshield now loads each config as a *dict*, merges the
dicts and creates a UserConfig instance from the merged dict. This fixes
the bug because a dict keeps the notion of whether a key was set or not,
so the merge works as expected.