fix: fix crash when scanning some commits
We would hit crashes on CI because of some very large line numbers (>2000
in a 80-lines patch).

Turns out `line_index_start` and `line_index_end` are relative to the
beginning of the document modified by the scanned patch, not relative to
the beginning of the patch itself.

With this commit, we no longer use these fields: instead we keep an
instance from MatchSpan in ExtendedMatch and use its fields.