refactor: remove the need for full content
In this commit we modify the output handlers to no longer depend on the
result.file.content, now the output entirely relies on the lines stored
in the extended match ! It also lead to discovering a number of issues
with the conftest and expected results in snap_test_text_output and
the cassetes.
We also added an additional test for Match span.
But for now the file is still stored in result, it should be removed in
the next commit.