fix: make sure `secret scan archive` does not overwrite files
`shutil.unpack_archive()` can overwrite files outside the destination
directory when unpacking a malicious tar archive. An attacker could use
this to overwrite files on the user machine.

This can only happen with tar archives, but the behavior with zip is not
optimal either: bad files are silently ignored.

With this change, the `secret scan archive` command now aborts if the
archive contains bad files or bad symbolic links, regardless of whether the
archive is a tar or a zip. For extra safety, the `secret scan pypi` command
does the same.