fix: fix git commit parser sometimes returning wrong diffs
Problem 1: sometimes the path for a diff is wrong. This is because
`commit_utils.parse_patch()` assumes that the diffs are in the same order
as the filenames in the commit header, which turns out to not always be
true.

Solution: get the path name from the `+++ b/path` or `--- a/path` lines of
the diff header.

Problem 2: turning merge commits into individual commits by calling
`git show` with the `-m` option does not work when fetching only some files
inside the commit (which we do when scanning huge commits): it always
return the changes from the first parent.

Solution: maybe it would be possible to return the changes from the correct
parents, but it felt simpler to stop using the `-m` option.
As a bonus, it simplifies `parse_patch()` because it can now assume that
there is only one commit per patch.