fix: fix parsing patches when the diff.noprefix git config option is set
Problem: if `diff.noprefix` is set in the git config, then our patch parser
fails.

Solution: set the appropriate environment variables to ignore any git
configuration. We still need a way to pass configuration because of the
"dubious ownership" error. This is what the GG_GIT_CONFIG environment
variable allows.

Note: We used to ignore the config in our test suite, but this commit
removes that to increase our chances of catching errors in case calls
to git which do not go through `git_shell.git()` are added in the future
(even if there should not be any).