refactor: rework verbose handling
Use ui.display_verbose() everywhere.

Remove many `verbose: bool` arguments.

Check if verbose is enabled using `ui.is_verbose()` instead of using
`config.user_config.verbose`. This avoids the need to update
`config.user_config.verbose` when called with `--verbose`, which is bad
because if the command rewrites the configuration file, it's going to
write `verbose: true` (this is still the case for `--allow-self-signed`).