chore: add a `get_data_dir()` function to core.dirs
This wraps `platformdirs.user_data_dir()`. It's going to be used as the
place to store our global git hook.

Changing the place of the global git hook is required to make the code
testable. It is also going to fix the fact that having the global git
hooks created in $HOME/.git/hooks was weird: for 99% of our users, $HOME is
not a git repository.