fix: fix build-os-packages failing when running inside the container
Problem: For some reason the ggshield ".git" dir is not visible inside
the Rocky Linux container when running inside GitHub CI. This causes
`build-os-packages --git-version` to fail because it can't run `git`.

Solution: Get the version suffix outside of the container and replace
the `--git-version` option with a `--suffix` option.