fix(with_incident_details): # We add a default value None to the click option to fix the behavior of the option
If no default value is defined, the behavior of ggshield can be misleading and actually bugged.
Indeed, each command or subcommand uses the decorator "add_secret_scan_common_options" to
collect values of several command options. This lets the user place the option --with-incident-details
at several level in the command line.

If they do so, the command or subcommand that does not have the flag will set the value to False
which can prevent the option from working correctly. For instance:
`ggshield secret scan --with-incident-details path dummy.py` => The option is set to False in the end because the path subcommand does not have the option defined.
`ggshield secret scan path --with-incident-details dummy.py` => The optiton is set to True, because defined in the path subcommand.

We make the option default to None. This does not override the value defined earlier in the command line.