fix: fix test_scan_context_repository()
Problem: Took me a while to understand why the test was failing. What
happened is the changes caused `secret scan path` to select more than one
file, and in this case it prompts for user confirmation, unless it's called
with `-y`. But the test does not report that because it does not test the
command result :(.

Solution:
1. This test does not need to scan multiple files, it just needs to be run
   inside a repository. Remove the `-r` arg and pass the path to the file
   to scan instead of passing the repository dir.
2. To make the test more robust, check the result of the scan call.
3. To keep things simple: do not include a secret in the test file, we
   don't care about finding secrets here.