fix(pre-commit): make import-linter-config hook work on Windows
The import-linter-config hook fails on Windows because calling a shell
script with `pdm run the_script` does not work on Windows: the shebang is
not taken into account.

Using `pdm run bash -c the_script` fix that, but then it fails because the
script itself calls ./scripts/generate-import-linter-config.py, but it does
not call it with the correct Python interpreter.

To fix that, call check-import-linter-config.sh without pdm, and make it
use `pdm run` to call generate-import-linter-config.py.