fix: Fix censor_string() behavior on short strings
When called with a one-char string, censor_string() would return the same
char doubled (so `censor_string("a")` would return "aa").

The behavior was also not good on 2-char string: it would leave the string
unchanged, and on 3-char string it would hide only the middle char.

Change this to make sure that:
- 1 & 2 char strings are fully censored
- 2 chars out of 3 are censored in a 3 char string

Fixes #1086